+++ phantom-dsl/src/main/scala/com/websudos/phantom/builder/query/CreateQueryOptions.scala
@@ -1,3 +1,32 @@
+/*
+ * Copyright 2013-2015 Websudos, Limited.
+ *
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * - Redistributions of source code must retain the above copyright notice,
+ * this list of conditions and the following disclaimer.
+ *
+ * - Redistributions in binary form must reproduce the above copyright
+ * notice, this list of conditions and the following disclaimer in the
+ * documentation and/or other materials provided with the distribution.
+ *
+ * - Explicit consent must be obtained from the copyright owner, Websudos Limited before any redistribution is made.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+ * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+ * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+ * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+ * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+ * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ */
 package com.websudos.phantom.builder.query
 
 import com.twitter.util.StorageUnit
@@ -16,13 +45,13 @@ sealed trait SpecifiedCompaction extends CompactionBound
 sealed trait UnspecifiedCompaction extends CompactionBound
 
 
-sealed class TablePropertyClause(val qb: CQLQuery) {}
+sealed class TablePropertyClause(val qb: CQLQuery)
 
 sealed abstract class CompactionStrategy(override val qb: CQLQuery) extends TablePropertyClause(qb)
 
 sealed trait CompactionStrategies {
 
-  private[this] def rootStrategy(strategy: String) = {
+  private[this] def strategy(strategy: String) = {
     CQLQuery(CQLSyntax.Symbols.`{`).forcePad
       .appendSingleQuote(CQLSyntax.CompactionOptions.`class`)
       .forcePad.append(CQLSyntax.Symbols.`:`)
@@ -31,11 +60,11 @@ sealed trait CompactionStrategies {
 
   sealed class SizeTieredCompactionStrategy(override val qb: CQLQuery) extends CompactionStrategy(qb) {
     def min_sstable_size(unit: StorageUnit): SizeTieredCompactionStrategy = {
-      new SizeTieredCompactionStrategy(QueryBuilder.Create.min_sstable_size(qb, unit.toHuman()))
+      new SizeTieredCompactionStrategy(QueryBuilder.Create.min_sstable_size(qb, unit.inMegabytes.toString))
     }
 
     def sstable_size_in_mb(unit: StorageUnit): SizeTieredCompactionStrategy = {
-      new SizeTieredCompactionStrategy(QueryBuilder.Create.sstable_size_in_mb(qb, unit.toHuman()))
+      new SizeTieredCompactionStrategy(QueryBuilder.Create.sstable_size_in_mb(qb, unit.inMegabytes.toString))
     }
 
     def bucket_high(size: Double): SizeTieredCompactionStrategy = {
@@ -47,9 +76,9 @@ sealed trait CompactionStrategies {
     }
   }
 
-  case object SizeTieredCompactionStrategy extends SizeTieredCompactionStrategy(rootStrategy(CQLSyntax.CompactionStrategies.SizeTieredCompactionStrategy))
-  case object LeveledCompactionStrategy extends CompactionStrategy(rootStrategy(CQLSyntax.CompactionStrategies.LeveledCompactionStrategy))
-  case object DateTieredCompactionStrategy extends CompactionStrategy(rootStrategy(CQLSyntax.CompactionStrategies.DateTieredCompactionStrategy))
+  case object SizeTieredCompactionStrategy extends SizeTieredCompactionStrategy(strategy(CQLSyntax.CompactionStrategies.SizeTieredCompactionStrategy))
+  case object LeveledCompactionStrategy extends CompactionStrategy(strategy(CQLSyntax.CompactionStrategies.LeveledCompactionStrategy))
+  case object DateTieredCompactionStrategy extends CompactionStrategy(strategy(CQLSyntax.CompactionStrategies.DateTieredCompactionStrategy))
 }
 
 sealed class CompressionStrategy(override val qb: CQLQuery) extends TablePropertyClause(qb) {
@@ -65,19 +94,19 @@ sealed class CompressionStrategy(override val qb: CQLQuery) extends TablePropert
 
 sealed trait CompressionStrategies {
 
-  private[this] def rootStrategy(strategy: String) = {
+  private[this] def strategy(strategy: String) = {
     CQLQuery(CQLSyntax.Symbols.`{`).forcePad
       .appendSingleQuote(CQLSyntax.CompressionOptions.sstable_compression)
       .forcePad.append(CQLSyntax.Symbols.`:`)
       .forcePad.appendSingleQuote(strategy)
   }
 
-  case object SnappyCompressor extends CompressionStrategy(rootStrategy(CQLSyntax.CompressionStrategies.SnappyCompressor))
-  case object LZ4Compressor extends CompressionStrategy(rootStrategy(CQLSyntax.CompressionStrategies.LZ4Compressor))
-  case object DeflateCompressor extends CompressionStrategy(rootStrategy(CQLSyntax.CompressionStrategies.DeflateCompressor))
+  case object SnappyCompressor extends CompressionStrategy(strategy(CQLSyntax.CompressionStrategies.SnappyCompressor))
+  case object LZ4Compressor extends CompressionStrategy(strategy(CQLSyntax.CompressionStrategies.LZ4Compressor))
+  case object DeflateCompressor extends CompressionStrategy(strategy(CQLSyntax.CompressionStrategies.DeflateCompressor))
 }
 
-sealed class CacheProperty(val qb: CQLQuery) {}
+sealed class CacheProperty(val qb: CQLQuery)
 
 object CacheStrategies {
   case object None extends CacheProperty(CQLQuery(CQLSyntax.CacheStrategies.None))
@@ -180,7 +209,7 @@ private[phantom] trait TablePropertyClauses extends CompactionStrategies with Co
       new TablePropertyClause(QueryBuilder.Create.replicate_on_write(clause.toString))
     }
 
-    def eqs = apply _
+    def eqs(clause: Boolean): TablePropertyClause = apply(clause)
   }
 
   object gc_grace_seconds extends TableProperty {