+++ phantom-dsl/src/test/scala/com/websudos/phantom/builder/serializers/AlterQueryBuilderTest.scala
@@ -112,7 +112,7 @@ class AlterQueryBuilderTest extends QueryBuilderTest {
         val qb = BasicTable.alter.`with`(compaction eqs SizeTieredCompactionStrategy.sstable_size_in_mb(50.megabytes)).qb.queryString
 
         qb shouldEqual "ALTER TABLE phantom.basicTable WITH compaction = { 'class' " +
-          ": 'SizeTieredCompactionStrategy', 'sstable_size_in_mb' : '50.0 MiB' }"
+          ": 'SizeTieredCompactionStrategy', 'sstable_size_in_mb' : '50' }"
       }
 
       "serialise a simple create query with a SizeTieredCompactionStrategy and 1 compaction strategy options set and a compression strategy set" in {
@@ -122,7 +122,7 @@ class AlterQueryBuilderTest extends QueryBuilderTest {
           .and(compression eqs LZ4Compressor.crc_check_chance(0.5))
           .qb.queryString
 
-        qb shouldEqual """ALTER TABLE phantom.basicTable WITH compaction = { 'class' : 'SizeTieredCompactionStrategy', 'sstable_size_in_mb' : '50.0 MiB' } AND compression = { 'sstable_compression' : 'LZ4Compressor', 'crc_check_chance' : 0.5 }"""
+        qb shouldEqual """ALTER TABLE phantom.basicTable WITH compaction = { 'class' : 'SizeTieredCompactionStrategy', 'sstable_size_in_mb' : '50' } AND compression = { 'sstable_compression' : 'LZ4Compressor', 'crc_check_chance' : 0.5 }"""
       }
 
       "add a comment option to a create query" in {