+++ phantom-dsl/src/test/scala/com/websudos/phantom/builder/serializers/CreateQueryBuilderTest.scala
@@ -65,7 +65,7 @@ class CreateQueryBuilderTest extends FreeSpec with Matchers with KeySpaceSuite {
         val qb = BasicTable.create.`with`(compaction eqs SizeTieredCompactionStrategy.sstable_size_in_mb(50.megabytes)).qb.queryString
 
         qb shouldEqual "CREATE TABLE phantom.basicTable (id uuid, id2 uuid, id3 uuid, placeholder text, PRIMARY KEY (id, id2, id3)) WITH compaction = { 'class' " +
-          ": 'SizeTieredCompactionStrategy', 'sstable_size_in_mb' : '50.0 MiB' }"
+          ": 'SizeTieredCompactionStrategy', 'sstable_size_in_mb' : '50' }"
       }
 
       "serialise a simple create query with a SizeTieredCompactionStrategy and 1 compaction strategy options set and a compression strategy set" in {
@@ -74,7 +74,7 @@ class CreateQueryBuilderTest extends FreeSpec with Matchers with KeySpaceSuite {
           .and(compression eqs LZ4Compressor.crc_check_chance(0.5))
         .qb.queryString
 
-        qb shouldEqual """CREATE TABLE phantom.basicTable (id uuid, id2 uuid, id3 uuid, placeholder text, PRIMARY KEY (id, id2, id3)) WITH compaction = { 'class' : 'SizeTieredCompactionStrategy', 'sstable_size_in_mb' : '50.0 MiB' } AND compression = { 'sstable_compression' : 'LZ4Compressor', 'crc_check_chance' : 0.5 }"""
+        qb shouldEqual """CREATE TABLE phantom.basicTable (id uuid, id2 uuid, id3 uuid, placeholder text, PRIMARY KEY (id, id2, id3)) WITH compaction = { 'class' : 'SizeTieredCompactionStrategy', 'sstable_size_in_mb' : '50' } AND compression = { 'sstable_compression' : 'LZ4Compressor', 'crc_check_chance' : 0.5 }"""
       }
 
       "add a comment option to a create query" in {